Add create cloud handler helper function
Currently when specifying a new cloud handler a user has to write
all of the request logic themselves. In some cases, such as wanting
only to append a new header to each request this is quite a lot of
work.

This adds a createCloudHandler function to fh-sync-js that allows
these situations to be handled in a nicer way. The request logic
itself remains in fh-sync-js.